refactor : 생성자 -> builder 패턴, exception handler 부분 수정
생성자를 그대로 쓰게 되면 필드가 추가되거나 변경되었을 때
에러 발생할 확률이 높고 손이 많이갈 확률 높음
롬복 @Builder 애용하기ㅋㅋ
ResponseEntity도 빌더 패턴 제공함.

아 그리고 id라는 말 막 사용하면 안되는게 엔티티 id랑 헷갈릴 수 있기 때문에 loginId로 명시하기 (id->loginId로 변경햇음)

에러 핸들러 부분에서 예외 메세지는 최대한 비즈니스 로직에서 직접 던지기 상황에 따라 다른 메세지가 올 수 있으니까
