--- src/main/java/brave/btc/exception/DuplicateIdException.java
@@ -1,4 +1,0 @@
-package brave.btc.exception;
-
-public class DuplicateIdException extends RuntimeException{
-}
