+++ src/main/java/brave/btc/exception/ErrorResponseDto.java
@@ -1,11 +1,16 @@
 package brave.btc.exception;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
 @AllArgsConstructor
-public class ErrorResult {
-    private String code;
+@Builder
+public class ErrorResponseDto<T>{
+
     private String message;
+    private T data;
 }
