+++ src/main/java/brave/btc/exception/auth/AuthenticationInvalidException.java
@@ -1,4 +1,4 @@
-package brave.btc.exception;
+package brave.btc.exception.auth;
 
 public class AuthenticationInvalidException extends RuntimeException{
