+++ src/main/java/brave/btc/exception/auth/DuplicateLoginIdException.java
@@ -1,0 +1,4 @@
+package brave.btc.exception.auth;
+
+public class DuplicateLoginIdException extends RuntimeException{
+}
