+++ src/main/java/brave/btc/exception/controller/ErrorController.java
@@ -1,11 +1,9 @@
 package brave.btc.exception.controller;
 
-import brave.btc.domain.User;
-import brave.btc.exception.AuthenticationInvalidException;
-import brave.btc.exception.ErrorResult;
-import jakarta.persistence.NoResultException;
+import brave.btc.exception.ErrorResponseDto;
+import brave.btc.exception.auth.AuthenticationInvalidException;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.EmptyResultDataAccessException;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -15,13 +13,25 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @RestControllerAdvice
 public class ErrorController {
 
-    @ExceptionHandler
-    public ResponseEntity<ErrorResult> invalidLoginExHandle(AuthenticationInvalidException e) {
-        log.error("[아이디/비밀번호 오류]", e);
-        ErrorResult errorResult = new ErrorResult("LOGIN-EX", e.getMessage());
-        return new ResponseEntity<>(errorResult, HttpStatus.BAD_REQUEST);
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponseDto<?>> handleException(Exception e) {
+
+        log.error("[handleException] 특정 되지 않은 예외");
+        ErrorResponseDto<Object> errorResponseDto = ErrorResponseDto.builder()
+            .message(e.getMessage())
+            .build();
+        return ResponseEntity.internalServerError()
+            .body(errorResponseDto);
     }
 
+    @ExceptionHandler(AuthenticationInvalidException.class)
+        public ResponseEntity<ErrorResponseDto<?>> handleAuthenticaionInvalidException(AuthenticationInvalidException e) {
 
-
+        log.error("[handleAuthenticationInvalidException] 로그인 실패 예외");
+        ErrorResponseDto<Object> errorResponseDto = ErrorResponseDto.builder()
+            .message(e.getMessage())
+            .build();
+        return ResponseEntity.badRequest()
+            .body(errorResponseDto);
+    }
 }
